Pigeon

Documentation for Pigeon.

Pigeon.isdominated

Pigeon.isimplied

Pigeon.sympermutations

Pigeon.sympermutations

Pigeon.isdominated — Method

isdominated(a, b)
Given abstract set expressions a and b, return true when b dominates a.
ArgumentError if the answer cannot be determined.

Pigeon.isimplied — Method

isimplied(a, b)
Given abstract predicate expressions a and b, return true when b implies a.
ArgumentError if the answer cannot be determined.

Pigeon.sympermutations — Method

sympermutations(u, v)

Generate all σ such that u[σ] == v and isunique(σ).

Pigeon.sympermutations — Method

sympermutations(v)

Generate all σ such that v[σ] == v and isunique(σ).